Consolidate admin panel handlers
